requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.7
django-filter==2.4.0
django-graphql-auth==0.3.15
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
promise==2.3
PyJWT==1.7.1
pytz==2021.1
regex==2020.11.13
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
